+++ pytorch_lightning/utilities/model_summary.py
@@ -23,8 +23,7 @@ from torch import Tensor
 from torch.utils.hooks import RemovableHandle
 
 import pytorch_lightning as pl
-from pytorch_lightning.utilities import AMPType, DeviceType, ModelSummaryMode, rank_zero_deprecation
-from pytorch_lightning.utilities.exceptions import MisconfigurationException
+from pytorch_lightning.utilities import AMPType, DeviceType
 from pytorch_lightning.utilities.imports import _TORCH_GREATER_EQUAL_1_8
 from pytorch_lightning.utilities.warnings import WarningCache
 
@@ -130,13 +129,6 @@ class ModelSummary:
 
     Args:
         model: The model to summarize (also referred to as the root module).
-        mode: Can be one of
-
-            - `top` (default): only the top-level modules will be recorded (the children of the root module)
-            - `full`: summarizes all layers and their submodules in the root module
-
-            .. deprecated:: v1.4
-                This parameter was deprecated in v1.4 in favor of `max_depth` and will be removed in v1.6.
 
         max_depth: Maximum depth of modules to show. Use -1 to show all modules or 0 to show no
             summary. Defaults to 1.
@@ -186,21 +178,8 @@ class ModelSummary:
         0.530     Total estimated model params size (MB)
     """
 
-    def __init__(self, model: "pl.LightningModule", mode: Optional[str] = None, max_depth: Optional[int] = 1) -> None:
+    def __init__(self, model: "pl.LightningModule", max_depth: int = 1) -> None:
         self._model = model
-
-        # temporary mapping from mode to max_depth
-        if max_depth is None or mode is not None:
-            if mode in ModelSummaryMode.supported_types():
-                max_depth = ModelSummaryMode.get_max_depth(mode)
-                rank_zero_deprecation(
-                    "Argument `mode` in `ModelSummary` is deprecated in v1.4"
-                    f" and will be removed in v1.6. Use `max_depth={max_depth}` to replicate `mode={mode}` behaviour."
-                )
-            else:
-                raise MisconfigurationException(
-                    f"`mode` can be {', '.join(ModelSummaryMode.supported_types())}, got {mode}."
-                )
 
         if not isinstance(max_depth, int) or max_depth < -1:
             raise ValueError(f"`max_depth` can be -1, 0 or > 0, got {max_depth}.")
@@ -436,17 +415,11 @@ def _is_lazy_weight_tensor(p: Tensor) -> bool:
     return False
 
 
-def summarize(
-    lightning_module: "pl.LightningModule", mode: Optional[str] = None, max_depth: Optional[int] = None
-) -> ModelSummary:
+def summarize(lightning_module: "pl.LightningModule", max_depth: int = 1) -> ModelSummary:
     """Summarize the LightningModule specified by `lightning_module`.
 
     Args:
         lightning_module: `LightningModule` to summarize.
-        mode: Can be either ``'top'`` (summarize only direct submodules) or ``'full'`` (summarize all layers).
-
-            .. deprecated:: v1.4
-                This parameter was deprecated in v1.4 in favor of `max_depth` and will be removed in v1.6.
 
         max_depth: The maximum depth of layer nesting that the summary will include. A value of 0 turns the
             layer summary off. Default: 1.
@@ -454,22 +427,4 @@ def summarize(
     Return:
         The model summary object
     """
-
-    # temporary mapping from mode to max_depth
-    if max_depth is None:
-        if mode is None:
-            model_summary = ModelSummary(lightning_module, max_depth=1)
-        elif mode in ModelSummaryMode.supported_types():
-            max_depth = ModelSummaryMode.get_max_depth(mode)
-            rank_zero_deprecation(
-                "Argument `mode` in `LightningModule.summarize` is deprecated in v1.4"
-                f" and will be removed in v1.6. Use `max_depth={max_depth}` to replicate `mode={mode}` behavior."
-            )
-            model_summary = ModelSummary(lightning_module, max_depth=max_depth)
-        else:
-            raise MisconfigurationException(
-                f"`mode` can be None, {', '.join(ModelSummaryMode.supported_types())}, got {mode}"
-            )
-    else:
-        model_summary = ModelSummary(lightning_module, max_depth=max_depth)
-    return model_summary
+    return ModelSummary(lightning_module, max_depth=max_depth)