+++ tests/deprecated_api/test_remove_1-6.py
@@ -18,7 +18,6 @@ import pytest
 
 from pytorch_lightning import Trainer
 from pytorch_lightning.utilities.model_helpers import is_overridden
-from pytorch_lightning.utilities.model_summary import ModelSummary
 from tests.helpers import BoringModel
 
 
@@ -67,15 +66,6 @@ def test_v1_6_0_is_overridden_model():
         assert is_overridden("validation_step", model=model)
     with pytest.deprecated_call(match="and will be removed in v1.6"):
         assert not is_overridden("foo", model=model)
-
-
-def test_v1_6_0_deprecated_model_summary_mode(tmpdir):
-    model = BoringModel()
-    with pytest.deprecated_call(match="Argument `mode` in `ModelSummary` is deprecated in v1.4"):
-        ModelSummary(model, mode="top")
-
-    with pytest.deprecated_call(match="Argument `mode` in `LightningModule.summarize` is deprecated in v1.4"):
-        model.summarize(mode="top")
 
 
 def test_v1_6_0_deprecated_disable_validation():